requirements.txt
discord
fastapi
python-dotenv
uvicorn
jinja2